ml-classifier/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
pymongo
numpy
python-dotenv
edge-tts